fix by review comments

+++ src/pstd/include/cache/config.h
@@ -23,9 +23,9 @@ enum RedisMaxmemoryPolicy {
   CACHE_NO_EVICTION       = 7
 };
 
-#define CACHE_DEFAULT_MAXMEMORY         (10 << 30) // 10G
+#define CACHE_DEFAULT_MAXMEMORY         ((uint64_t)(10) << 30) // 10G
 #define CACHE_DEFAULT_MAXMEMORY_SAMPLES 5
-#define CACHE_DEFAULT_LFU_DECAY_TIME    1 
+#define CACHE_DEFAULT_LFU_DECAY_TIME    1
 
 /*
  * cache start pos
